requirements.txt
absl-py==0.11.0
av==8.0.2
cachetools==4.2.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cycler==0.10.0
dataclasses==0.8
decorator==4.4.2
einops==0.3.0
Flask==1.1.2
google-auth==1.24.0
google-auth-oauthlib==0.4.2
grpcio==1.34.1
idna==2.10
imageio==2.9.0
importlib-metadata==3.4.0
itsdangerous==1.1.0
Jinja2==2.11.3
kiwisolver==1.3.1
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.4
networkx==2.5
numpy==1.19.5
oauthlib==3.1.0
omegaconf==2.0.5
opencv-python==4.5.1.48
pandas==1.1.5
Pillow==8.1.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.5
PyWavelets==1.1.1
PyYAML==5.4b2
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7
scikit-image==0.17.2
scikit-video==1.1.11
scipy==1.5.4
six==1.15.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.7.0
tifffile==2020.9.3
torch==1.7.1
torchsummary==1.5.1
torchvision==0.8.2
typing-extensions==3.7.4.3
urllib3==1.26.2
videoaugment @ git+https://github.com/iliasprc/videoaugment@c2a15b36f4fb22683e865405f204c7c4a4fc5ecc
Werkzeug==1.0.1
zipp==3.4.0
